+++ nix/haskell-tooling-overlay.nix
@@ -1,0 +1,7 @@
+{ self }:
+final: prev: {
+  haskell-tools-nvim-dev = prev.vimUtils.buildVimPluginFrom2Nix {
+    name = "haskell-tools-nvim";
+    src = self;
+  };
+}